chore: cleanup and added the checkout link in home
